+++ main.go
@@ -4,92 +4,65 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"net/http"
 	"os"
-	"os/signal"
-	"sync"
-	"syscall"
 	"time"
 
-	log "github.com/sirupsen/logrus"
+	secretsmanagerv1alpha1 "github.com/tuenti/secrets-manager/api/v1alpha1"
 	"github.com/tuenti/secrets-manager/backend"
-	k8s "github.com/tuenti/secrets-manager/kubernetes"
-	secretsmanager "github.com/tuenti/secrets-manager/secrets-manager"
-
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-
-	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"github.com/tuenti/secrets-manager/controllers"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/log/zap"
+	// +kubebuilder:scaffold:imports
 )
+
+var (
+	scheme   = runtime.NewScheme()
+	setupLog = ctrl.Log.WithName("setup")
+)
+
+func init() {
+	corev1.AddToScheme(scheme)
+	secretsmanagerv1alpha1.AddToScheme(scheme)
+	// +kubebuilder:scaffold:scheme
+}
 
 // To be filled from build ldflags
 var version string
 
-func newK8sClientSet() (*kubernetes.Clientset, error) {
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		return nil, err
-	}
-
-	clientSet, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-	return clientSet, nil
-}
-
 func main() {
-	var logger *log.Logger
-	var wg sync.WaitGroup
-
+	var metricsAddr string
+	var enableLeaderElection bool
+	var enableDebugLog bool
+	var versionFlag bool
+	var reconcilePeriod time.Duration
+	var selectedBackend string
 	backendCfg := backend.Config{}
-	secretsManagerCfg := secretsmanager.Config{}
-	selectedBackend := flag.String("backend", "vault", "Selected backend. Only vault supported")
-	logLevel := flag.String("log.level", "warn", "Minimum log level")
-	logFormat := flag.String("log.format", "text", "Log format, one of text or json")
-	versionFlag := flag.Bool("version", false, "Display Secret Manager version")
-	addr := flag.String("listen-address", ":8080", "The address to listen on for HTTP requests.")
-
-	flag.StringVar(&secretsManagerCfg.ConfigMap, "config.config-map", "secrets-manager-config", "Name of the config Map with Secrets Manager settings (format: [<namespace>/]<name>) ")
-	flag.DurationVar(&secretsManagerCfg.BackendScrapeInterval, "config.backend-scrape-interval", 15*time.Second, "Scraping secrets from backend interval")
-	flag.DurationVar(&secretsManagerCfg.ConfigMapRefreshInterval, "config.configmap-refresh-interval", 15*time.Second, "ConfigMap refresh interval")
-
+	flag.StringVar(&metricsAddr, "metrics-addr", ":8080", "The address the metric endpoint binds to.")
+	flag.BoolVar(&enableLeaderElection, "enable-leader-election", false,
+		"Enable leader election for controller manager. Enabling this will ensure there is only one active controller manager.")
+	flag.StringVar(&selectedBackend, "backend", "vault", "Selected backend. Only vault supported")
+	flag.BoolVar(&enableDebugLog, "enable-debug-log", false, "Enable this to get more logs verbosity and debug messages.")
+	flag.BoolVar(&versionFlag, "version", false, "Display Secret Manager version")
+	flag.DurationVar(&reconcilePeriod, "reconcile-period", 5*time.Second, "How often the controller will re-queue secretdefinition events")
 	flag.DurationVar(&backendCfg.BackendTimeout, "config.backend-timeout", 5*time.Second, "Backend connection timeout")
 	flag.StringVar(&backendCfg.VaultURL, "vault.url", "https://127.0.0.1:8200", "Vault address. VAULT_ADDR environment would take precedence.")
 	flag.StringVar(&backendCfg.VaultRoleID, "vault.role-id", "", "Vault approle role id. VAULT_ROLE_ID environment would take precedence.")
-	flag.StringVar(&backendCfg.VaultRoleID, "vault.secret-id", "", "Vault approle secret id. VAULT_SECRET_ID environment would take precedence.")
+	flag.StringVar(&backendCfg.VaultSecretID, "vault.secret-id", "", "Vault approle secret id. VAULT_SECRET_ID environment would take precedence.")
 	flag.Int64Var(&backendCfg.VaultMaxTokenTTL, "vault.max-token-ttl", 300, "Max seconds to consider a token expired.")
 	flag.DurationVar(&backendCfg.VaultTokenPollingPeriod, "vault.token-polling-period", 15*time.Second, "Polling interval to check token expiration time.")
 	flag.IntVar(&backendCfg.VaultRenewTTLIncrement, "vault.renew-ttl-increment", 600, "TTL time for renewed token.")
 	flag.StringVar(&backendCfg.VaultEngine, "vault.engine", "kv2", "Vault secret engine. Only KV version 1 and 2 supported")
 	flag.Parse()
 
-	if *versionFlag {
+	if versionFlag {
 		fmt.Printf("Secrets Manager %s\n", version)
 		os.Exit(0)
 	}
 
-	logger = log.New()
-
-	switch *logLevel {
-	case "info":
-		logger.SetLevel(log.InfoLevel)
-	case "err":
-		logger.SetLevel(log.ErrorLevel)
-	case "debug":
-		logger.SetLevel(log.DebugLevel)
-	default:
-		logger.SetLevel(log.WarnLevel)
-	}
-
-	switch *logFormat {
-	case "json":
-		logger.Formatter = &log.JSONFormatter{}
-	default:
-		logger.Formatter = &log.TextFormatter{}
-	}
-
-	logger.SetOutput(os.Stdout)
+	logger := zap.Logger(enableDebugLog).WithName("backend")
 
 	if os.Getenv("VAULT_ADDR") != "" {
 		backendCfg.VaultURL = os.Getenv("VAULT_ADDR")
@@ -106,72 +79,41 @@ func main() {
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	backendClient, err := backend.NewBackendClient(ctx, *selectedBackend, logger, backendCfg)
+	backendClient, err := backend.NewBackendClient(ctx, selectedBackend, logger, backendCfg)
 	if err != nil {
-		logger.Errorf("could not build backend client: %v", err)
+		logger.Error(err, "could not build backend client")
 		os.Exit(1)
 	}
 
-	clientSet, err := newK8sClientSet()
+	ctrl.SetLogger(zap.Logger(enableDebugLog))
 
+	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
+		Scheme:             scheme,
+		MetricsBindAddress: metricsAddr,
+		LeaderElection:     enableLeaderElection,
+	})
 	if err != nil {
-		logger.Errorf("could not build k8s client: %v", err)
+		setupLog.Error(err, "unable to start manager")
 		os.Exit(1)
 	}
 
-	kubernetes := k8s.New(clientSet, logger)
-	secretsManager, err := secretsmanager.New(ctx, secretsManagerCfg, kubernetes, *backendClient, logger)
-
+	err = (&controllers.SecretDefinitionReconciler{
+		Backend:              *backendClient,
+		Client:               mgr.GetClient(),
+		APIReader:            mgr.GetAPIReader(),
+		Log:                  ctrl.Log.WithName("controllers").WithName("SecretDefinition"),
+		Ctx:                  ctx,
+		ReconciliationPeriod: reconcilePeriod,
+	}).SetupWithManager(mgr)
 	if err != nil {
-		logger.Errorf("could not init Secret Manager: %v", err)
+		setupLog.Error(err, "unable to create controller", "controller", "SecretDefinition")
 		os.Exit(1)
 	}
+	// +kubebuilder:scaffold:builder
 
-	sigc := make(chan os.Signal, 1)
-	signal.Notify(sigc,
-		syscall.SIGHUP,
-		syscall.SIGTERM,
-		syscall.SIGQUIT)
-
-	wg.Add(1)
-	go secretsManager.Start(ctx)
-
-	srv := startHttpServer(*addr, logger)
-
-	for {
-		select {
-		case <-sigc:
-			shutdownHttpServer(srv, logger)
-			cancel()
-			break
-		}
-		break
+	setupLog.Info("starting manager")
+	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {
+		setupLog.Error(err, "problem running manager")
+		os.Exit(1)
 	}
-	wg.Wait()
-}
-
-func shutdownHttpServer(srv *http.Server, logger *log.Logger) {
-	logger.Infof("[main] Stopping HTTP server")
-
-	if err := srv.Shutdown(nil); err != nil {
-		logger.Errorf("ListenAndServe(): %s", err)
-	} else {
-		logger.Infof("[main] Stopped HTTP server")
-	}
-}
-
-func startHttpServer(addr string, logger *log.Logger) *http.Server {
-	srv := &http.Server{Addr: addr}
-
-	http.Handle("/metrics", promhttp.Handler())
-
-	go func() {
-		logger.Infof("Starting HTTP server listening on %v", addr)
-		// returns ErrServerClosed on graceful close
-		if err := srv.ListenAndServe(); err != http.ErrServerClosed {
-			logger.Errorf("[main] Unexpected error in HTTP server: %s", err)
-		}
-	}()
-
-	return srv
 }